+++ src/compojure/api/coercion/spec.clj
@@ -6,7 +6,9 @@
             [clojure.walk :as walk]
             [compojure.api.coercion.core :as cc]
             [spec-tools.swagger.core :as swagger]
-            [compojure.api.common :as common])
+            [compojure.api.common :as common]
+            ;; side effects
+            compojure.api.coercion.register-spec)
   (:import (clojure.lang IPersistentMap)
            (schema.core RequiredKey OptionalKey)
            (spec_tools.core Spec)
@@ -149,5 +151,3 @@
   (->SpecCoercion :spec options))
 
 (def default-coercion (create-coercion default-options))
-
-(defmethod cc/named-coercion :spec [_] default-coercion)